+++ src/app/register/page.tsx
@@ -1,12 +1,5 @@
 "use client";
 
-import { Button, Header, Input, Menu } from "@/components";
-import Image from "next/image";
-
-export default function Login() {
-  return (
-    <div className="w-full flex justify-center items-center">
-      <Header />
-    </div>
-  );
+export default function Register() {
+  return <div className="w-full flex justify-center items-center"></div>;
 }
